Refactor code to avoid circular import
